@@ -2,7 +2,7 @@ package iotdevice.com.iotDevice.common
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import iotdevice.com.iotDevice.member.AuthenticationServiceLocator
 
 abstract class BaseActivity : AppCompatActivity() {
